backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-docx>=1.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
python-multipart>=0.0.9
